altera tipo de dado para double, a fim de minimizar casos de erro

+++ global.c
@@ -1,6 +1,6 @@
-float *vetor = NULL; //vetor de entrada
-float somaThreads;
-float somaEsperada;
+double *vetor = NULL; //vetor de entrada
+double somaThreads;
+double somaEsperada;
 
 typedef struct {
     int id; //id da thread
